backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
SQLAlchemy
python-dotenv
pydantic
alembic
email-validator